Added code for formatting tree widget menu, and graph dialog box
Just added some of things I was messing around with the passed week.
Additions include: Tree Widget items (just an example of 3 layers), and
a graph example on a new dialog window.

+ I could not figure out how to get Graphs file tab to enable itself,
so I put a hot key in of “ctrl+g”  to open the graph dialog window
(just for now)